+++ CHANGELOG.md
@@ -51,6 +51,8 @@
 - Improve performance when calling built-in functions from the Python scope
   ([GH-801](https://github.com/NVIDIA/warp/issues/801)).
 - Building `warp.fem` geometry and function space partitions is now possible in CUDA graphs by passing an explicit upper-bound for the number of cells and nodes to `ExplicitGeometryPartition` and `make_space_partition`. Additionally, building fields and field restrictions is now synchronization-free by default ([GH-1021](https://github.com/NVIDIA/warp/issues/1021)).
+- Raise `RuntimeError` from `wp.load_module()` when attempting to load a module that does not contain any Warp kernels,
+  functions, or structs ([GH-920](https://github.com/NVIDIA/warp/issues/920)).
 
 
 ### Fixed